Prioritize results by number of direct matches
Before trying a full indirect regex search (with gaps allowed),
try to do a search without gaps allowed. Display results with
more direct matches first.

Note: this is not the same as "fewest gaps", but that kind of
optimizing search would not be possible with regular expressions.
This solution should be a satisfying replacement.

Fixes #3.